backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-multipart
passlib[bcrypt]
python-jose
stripe
httpx
